# backend/scripts/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Campus Operations AI API",
                 "description":  "API for Campus Operations Intelligence Platform",
                 "version":  "0.0.1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/recommendations":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "recommendation-controller"
                                                                    ],
                                                           "summary":  "Recommendations for all schedules",
                                                           "description":  "Return recommendations for all schedules",
                                                           "operationId":  "all",
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  "@{schema=}"
                                                                                                     }
                                                                                     }
                                                                         },
                                                           "security":  [
                                                                            {
                                                                                "bearerAuth":  [

                                                                                               ]
                                                                            }
                                                                        ]
                                                       }
                                           },
                  "/api/recommendations/room/{scheduleId}":  {
                                                                 "get":  {
                                                                             "tags":  [
                                                                                          "recommendation-controller"
                                                                                      ],
                                                                             "summary":  "Recommend room for schedule",
                                                                             "description":  "Recommend a better room for the given schedule id",
                                                                             "operationId":  "recommendRoom",
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "scheduleId",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int64"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "OK",
                                                                                                           "content":  {
                                                                                                                           "*/*":  "@{schema=}"
                                                                                                                       }
                                                                                                       }
                                                                                           },
                                                                             "security":  [
                                                                                              {
                                                                                                  "bearerAuth":  [

                                                                                                                 ]
                                                                                              }
                                                                                          ]
                                                                         }
                                                             },
                  "/api/dashboard/summary":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "dashboard-controller"
                                                                      ],
                                                             "summary":  "Dashboard summary",
                                                             "description":  "Counts of students, rooms, courses, and schedules",
                                                             "operationId":  "summary",
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  "@{schema=}"
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                             },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}